Clear the tmux history before running tests anew
When I run tests in a tmux pane, I like being able to use `[` to go to
the first line and page around. To give a clean buffer and make it clear
that all output belongs to the current run, I should just clobber the
most recent ouput from the last run.