Delete trailing whitespace on Markdown files
I have Vim configured to _always_ strip trailing whitespace for a couple
of file extensions. Generally, when working in Git, a requirement for no
trailing whitespace can be controlled at the repository level. However,
I write a lot of Markdown outside of Git repositories. Apply that
automatic stripping as well.