Automatically resource tmuxrc after editing
NB: The ONLY way to test changes to a .tmux.conf is to restart tmux.

This is because a current session will still honor the settings that
were invoked previously. As an example, add the following to your
`.tmux.conf`:

    set -g status-justify centre

Observe that the status bar is now centered.

Delete that line, and resource `.tmux.conf`. Notice that the status bar
remains centered, since the default tmux settings are not re-sourced.

So, in order to fully figure out the changes that will result from a new
.tmuxrc, one actually has to restart all of `tmux`. The tmux developers
have weighed in on this, and they advocate a hard restart. Tmux devs
prefer to not build a "setting refresh" option that restores all
defaults. There are "Tmux reset" tools out there that function like a
CSS reset (explicitly defining all defaults), but this is not likely to
scale well with continued Tmux development.