Annotate how to use `vim-instant-markdown`
This package requires a global installation of a special daemon. It's
only really useful when using Vim within a desktop environment, so I
won't auto-install this daemon.

The 'for markdown' addition to `Plug` isn't strictly necessary, but it
does make it clear that this plugin is only needed when editing Markdown
files.

The setting of `autostart` behavior is also not strictly necessary,
since that's the default. However, it gives quick reference to my future
self about how to disable automatic daemon + browser starting, should
this start to annoy me.