Add shortcuts for quick & easy copy of GitHub refs
When discussing code with peers, I'll very often want a shareable link
to just a single line, or to a selection of lines.

`:Gbrowse!` copies the link, instead of opening a browser like
`:Gbrowse` does.

However, `:Gbrowse!` will (unless in visual mode), just copy a reference
to the whole file. I almost always want to reference lines. `,c` will
copy a reference to just the current line or selection.