Remove vim-prettier as a plugin
Normally, using `ale` where the `prettier` binary is accessible through
`node_modules` is plenty to auto-detect prettier for linting (and it's
easier still to set it up as an auto-linter).

In some rare cases, I enjoy being able to run `:Prettier` on a file
where ALE is *not* set up (e.g. to auto-format a JSON file or similar).
However, I can always just use the `prettier` binary for those rare
cases.